get random cities

+++ src/main/java/ch/ttt/cityguessrservice/cities/City.java
@@ -3,11 +3,8 @@ package ch.ttt.cityguessrservice.cities;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Value;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.util.UUID;
 
 @Document
 @Data
@@ -15,7 +12,7 @@ import java.util.UUID;
 @AllArgsConstructor
 public class City {
     @Id
-    private Integer id;
+    private Long id;
     private String name;
     private String lat;
     private String lng;
